Changed SOCRATES wrapper for optionally avoiding extra fileIO and avoiding creating the rscatter spectral file each call. RCE achieved in approx 5 minutes per run.

+++ utils/SocRadModel.py
@@ -18,7 +18,8 @@ from utils.atmosphere_column import atmos
 import utils.phys as phys
 
 
-def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True):
+def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,
+               rewrite_cfg=True, rewrite_tmp=True, rewrite_gas=True):
     """Runs SOCRATES to calculate fluxes and heating rates
 
     Parameters
@@ -29,10 +30,17 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
             Named directories
         recalc : bool
             Is this function call a 'recalculation' case accounting for a tropopause?
+            
         calc_cf : bool
             Calculate contribution function?
         rscatter : bool
             Include Rayleigh scattering?
+        write_cfg : bool
+            Re-write configuration values (e.g. TOA heating)
+        write_PT : bool
+            Re-write temperature and pressure arrays
+        write_gas : bool
+            Re-write composition files
             
     """
 
@@ -80,70 +88,75 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
     
     # Rayleigh scattering for CO2
     if rscatter == True:
-
-        scatter_dir = dirs["output"]
-
+        
         # New file
         spectral_file_old = spectral_file
-        spectral_file = scatter_dir+"runtime_spectral_file_rscat"
+        spectral_file = dirs["output"]+"runtime_spectral_file_rscat"
 
-        shutil.copyfile(spectral_file_old,spectral_file)
-        shutil.copyfile(spectral_file_old+"_k",spectral_file+"_k")
+        # Skip if already exists
+        if (not os.path.exists(spectral_file)) or rewrite_gas:
 
-        # Insert Rayleigh scattering into spectral file
-        RayleighSpectrum.rayleigh_coeff_adder(species_list = ['co2', 'h2o', 'n2'], 
-                                              mixing_ratio_list = [atm.x_gas["CO2"][-1], atm.x_gas["H2O"][-1], atm.x_gas["N2"][-1]], 
-                                              spectral_file_path=spectral_file,
-                                              wavelength_dummy_file_path=dirs["output"]+'wavelength_band_file.txt'
-                                              )
+            shutil.copyfile(spectral_file_old,spectral_file)
+            shutil.copyfile(spectral_file_old+"_k",spectral_file+"_k")
 
+            # Insert Rayleigh scattering into spectral file
+            RayleighSpectrum.rayleigh_coeff_adder(species_list = ['co2', 'h2o', 'n2'], 
+                                                mixing_ratio_list = [atm.x_gas["CO2"][-1], atm.x_gas["H2O"][-1], atm.x_gas["N2"][-1]], 
+                                                spectral_file_path=spectral_file,
+                                                wavelength_dummy_file_path=dirs["output"]+'wavelength_band_file.txt'
+                                                )
         scatter_flag = " -r"
     else:
         scatter_flag = ""
 
 
-    # # Enable cf SOCRATES environment
-    # if calc_cf == True:
-    #     path_to_cff_socrates = dirs["rad_conv"]+"/rad_trans/socrates_cff"
-    #     call_sequence = [ "source", path_to_cff_socrates+"/set_rad_env" ]
-    #     print(call_sequence)
-    #     subprocess.call(call_sequence)
-
-    # Remove auxiliary files from previous runs
-    CleanOutputDir( os.getcwd() )
-
-    # Solar zenith angle
-    zenith_angle    = atm.zenith_angle  # Hamano+15: 54.7, Ranjan+18: 48.2, Katyal+19: 38
-
-    # Surface albedo
-    surface_albedo  = atm.albedo_s      # Hamano+15: 0.2, Schaefer+16: 0.75
-
-    # Other parameters
-    longitude       = 0
-    latitude        = 0
-    basis_function  = 1
-
+    # Write values to netcdf: SOCRATES Userguide p. 45
     basename = 'profile'
 
-    # Write values to netcdf: SOCRATES Userguide p. 45
-    nctools.ncout_surf(basename+'.surf', longitude, latitude, basis_function, surface_albedo)
-    nctools.ncout2d(   basename+'.tstar', 0, 0, atm.ts, 'tstar', longname="Surface Temperature", units='K')
-    nctools.ncout2d(   basename+'.pstar', 0, 0, atm.ps, 'pstar', longname="Surface Pressure", units='PA')
-    nctools.ncout2d(   basename+'.szen', 0, 0, zenith_angle, 'szen', longname="Solar zenith angle", units='Degrees')
-    nctools.ncout2d(   basename+'.stoa', 0, 0, atm.toa_heating, 'stoa', longname="Solar Irradiance at TOA", units='WM-2')
+    # Write configuration stuff
+    check_cfg = lambda fpath: rewrite_cfg or (not os.path.exists(fpath))
 
+    fthis = basename+'.surf'
+    if check_cfg(fthis): nctools.ncout_surf(fthis, 0, 0, 1, float(atm.albedo_s))
+
+    fthis = basename+'.tstar'
+    if check_cfg(fthis): nctools.ncout2d(   fthis, 0, 0, atm.ts, 'tstar', longname="Surface Temperature", units='K')
+
+    fthis = basename+'.pstar'
+    if check_cfg(fthis): nctools.ncout2d(   fthis, 0, 0, atm.ps, 'pstar', longname="Surface Pressure", units='PA')
+    
+    fthis = basename+'.szen'
+    if check_cfg(fthis): nctools.ncout2d(   fthis, 0, 0, float(atm.zenith_angle), 'szen', longname="Solar zenith angle", units='Degrees')
+    
+    fthis = basename+'.stoa'
+    if check_cfg(fthis): nctools.ncout2d(   fthis, 0, 0, atm.toa_heating, 'stoa', longname="Solar Irradiance at TOA", units='WM-2')
 
     # T, P + volatiles
-    nctools.ncout3d(basename+'.t', 0, 0,   atm.p,  atm.tmp, 't', longname="Temperature", units='K')
-    nctools.ncout3d(basename+'.tl', 0, 0,  atm.pl, atm.tmpl, 'tl', longname="Temperature", units='K')
-    nctools.ncout3d(basename+'.p', 0, 0,   atm.p,  atm.p, 'p', longname="Pressure", units='PA')
-    nctools.ncout3d(basename+'.q', 0, 0,   atm.p,  molar_mass['H2O'] / atm.mu * atm.x_gas["H2O"], 'q', longname="q", units='kg/kg') 
+    check_tmp = lambda fpath: rewrite_tmp or (not os.path.exists(fpath))
+
+    fthis = basename+'.t'
+    if check_tmp(fthis): nctools.ncout3d(   fthis, 0, 0,   atm.p,  atm.tmp, 't', longname="Temperature", units='K')
+
+    fthis = basename+'.tl'
+    if check_tmp(fthis): nctools.ncout3d(   fthis, 0, 0,  atm.pl, atm.tmpl, 'tl', longname="Temperature", units='K')
+
+    fthis = basename+'.p'
+    if check_tmp(fthis): nctools.ncout3d(   fthis, 0, 0,   atm.p,  atm.p, 'p', longname="Pressure", units='PA')
+
+
+    # Gases
+    check_gas = lambda fpath: rewrite_gas or (not os.path.exists(fpath))
+
+    fthis = basename+'.q'
+    if check_gas(fthis): nctools.ncout3d(   fthis, 0, 0,   atm.p,  molar_mass['H2O'] / atm.mu * atm.x_gas["H2O"], 'q', longname="q", units='kg/kg') 
 
     allowed_vols = {"CO2", "O3", "N2O", "CO", "CH4", "O2", "NO", "SO2", "NO2", "NH3", "HNO3", "N2", "H2", "He", "OCS"}
     for vol in atm.vol_list.keys():
         if vol in allowed_vols:
             vol_lower = str(vol).lower()
-            nctools.ncout3d(basename+'.'+vol_lower, 0, 0, atm.p,  molar_mass[vol] / atm.mu * atm.x_gas[vol], vol_lower, longname=vol, units='kg/kg') 
+            fthis = basename+'.'+vol_lower
+            if check_gas(fthis):
+                nctools.ncout3d(basename+'.'+vol_lower, 0, 0, atm.p,  molar_mass[vol] / atm.mu * atm.x_gas[vol], vol_lower, longname=vol, units='kg/kg') 
 
     # Call sequences for run SOCRATES + move data
     seq_sw_ex = ["Cl_run_cdf","-B", basename,"-s", spectral_file, "-R 1", str(atm.nbands), " -ch ", str(atm.nbands), " -S -g 2 -C 5 -u", scatter_flag]
@@ -158,9 +171,10 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
 
     # Write namelist file?
     if socrates_use_namelist:
+
         mmw = 0.0
         for vol in atm.vol_list.keys():
-            mmw += ( np.mean(atm.x_gas[vol]) *  molar_mass[vol])   # There's probably a more accurate way to handle this.
+            mmw += ( np.mean(atm.x_gas[vol]) *  molar_mass[vol])  # Should be mass-weighted mean?
 
         rgas = phys.R_gas/mmw
         cp_avg = np.mean(atm.cp) / mmw
@@ -175,10 +189,13 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
             }
         }
 
-        f90nml.write(nml,basename+".nml")
+        fthis = basename+".nml"
+        if os.path.exists(fthis):
+            os.remove(fthis)
+        f90nml.write(nml,fthis)
         
-        seq_sw_ex.extend(["-N", basename+".nml"])
-        seq_lw_ex.extend(["-N", basename+".nml"])
+        seq_sw_ex.extend(["-N", fthis])
+        seq_lw_ex.extend(["-N", fthis])
 
     # Socrates print to stdout at runtime?
     if socrates_print == True:
@@ -206,27 +223,24 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
     ncfile2  = net.Dataset('currentsw.sflx')  # direct
     ncfile3  = net.Dataset('currentsw.dflx')  # diffuse 
     ncfile4  = net.Dataset('currentsw.uflx')  # upward
-    ncfile5  = net.Dataset('currentsw.nflx')  # net
     ncfile6  = net.Dataset('currentsw.hrts')  # heating rate
 
     ncfile7  = net.Dataset('currentlw.dflx')  # diffuse
-    ncfile8  = net.Dataset('currentlw.nflx')  # net
     ncfile9  = net.Dataset('currentlw.uflx')  # upward
     ncfile10 = net.Dataset('currentlw.hrts')  # heating rate
+
     if calc_cf == True:
         ncfile11 = net.Dataset('currentlw_cff.cff')
         ncfile12 = net.Dataset('currentlw.cff')
 
-
     # Loop through netCDF variables and populate arrays
     vflxsw   = ncfile1.variables['vflx']  # SW downward flux (direct + diffuse)
     uflxsw   = ncfile4.variables['uflx']  # SW upward flux 
-    nflxsw   = ncfile5.variables['nflx']  # SW net flux 
     hrtssw   = ncfile6.variables['hrts']  # SW heating rate (K/day)
     dflxlw   = ncfile7.variables['dflx']  # LW downward flux (diffuse)
     uflxlw   = ncfile9.variables['uflx']  # LW upward flux 
-    nflxlw   = ncfile8.variables['nflx']  # LW net flux 
     hrtslw   = ncfile10.variables['hrts'] # LW heating rate (K/day)
+
     if calc_cf == True:
         cff    = ncfile11.variables['cff'] # Contribution function (flux, sum)
         cff_i  = ncfile12.variables['cff'] # Contribution function (channel, plev, lat, lon)
@@ -285,10 +299,8 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
     ncfile2.close()
     ncfile3.close()
     ncfile4.close()
-    ncfile5.close()
     ncfile6.close()
     ncfile7.close()
-    ncfile8.close()
     ncfile9.close()
     ncfile10.close()
     if calc_cf == True:
@@ -296,7 +308,7 @@ def radCompSoc(atm, dirs, recalc, calc_cf=False, rscatter=False,insert_star=True
         # ncfile12.close()
 
     # Remove auxiliary files
-    CleanOutputDir( os.getcwd() )
+    # CleanOutputDir( os.getcwd() )
 
     return atm
 
@@ -314,11 +326,9 @@ def CleanOutputDir( output_dir ):
     for files in types:
         files_to_delete.extend(glob.glob(output_dir+"/"+files))
 
-    # print("Remove old output files:")
     for file in natural_sort(files_to_delete):
         os.remove(file)
-    #     print(os.path.basename(file), end =" ")
-    # print("\n==> Done.")
+
 
 # Disable and enable print: https://stackoverflow.com/questions/8391411/suppress-calls-to-print-python
 def blockPrint():
